home screen with restaurant list and dark mode toggle

+++ app/src/main/java/com/example/restaurantapp/data/local/entity/FavoriteRestaurantEntity.kt
@@ -26,6 +26,6 @@ class FavoriteRestaurantEntity(
     val city: String,
 
     @field:ColumnInfo(name = "rating")
-    val rating: Float
+    val rating: Double,
 ): Parcelable
